--- Content.Omu.Client/LightDetection/ShadowkinLightDetectionSystem.cs
@@ -1,5 +1,0 @@
-﻿using Content.Omu.Shared.ShadowkinLightDetection.Systems;
-
-namespace Content.Omu.Client.ShadowkinLightDetection;
-
-public sealed class ShadowkinLightDetectionSystem : SharedShadowkinLightDetectionSystem;
