--- Content.Omu.Shared/LightDetection/Systems/SharedShadowkinLightDetectionSystem.cs
@@ -1,3 +1,0 @@
-﻿namespace Content.Omu.Shared.ShadowkinLightDetection.Systems;
-
-public abstract class SharedShadowkinLightDetectionSystem : EntitySystem;
