+++ Content.Omu.Shared/Species/Shadowkin/LightDetection/Components/ShadowkinLightDetectionDamageComponent.cs
@@ -3,10 +3,9 @@ using Content.Shared.Damage;
 using Robust.Shared.Audio;
 using Robust.Shared.GameStates;
 using Robust.Shared.Prototypes;
-using Robust.Shared.Serialization.TypeSerializers.Implementations;
 using Robust.Shared.Serialization.TypeSerializers.Implementations.Custom;
 
-namespace Content.Omu.Shared.ShadowkinLightDetection.Components;
+namespace Content.Omu.Shared.Species.Shadowkin.LightDetection.Components;
 
 /// <summary>
 /// Component that indicates a user should take damage or heal damage based on the light detection system
